pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "research-harness"
version = "0.1.0"
description = "Agent-to-agent conversation loop for studying emergent behavior"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "anthropic>=0.40",
    "google-genai>=1.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "pydantic>=2.0",
]

[project.scripts]
rh = "research_harness.cli:main"
